+++ africanus/coordinates/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+from .coordinates import radec_to_lmn, lmn_to_radec