+++ africanus/gridding/simple/dask.py
@@ -4,6 +4,9 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+from functools import reduce
+from operator import mul
+
 from .gridding import (grid as np_grid_fn, degrid as np_degrid_fn)
 from ...util.docs import on_rtd, mod_docs
 from ...util.requirements import have_packages, MissingPackageException
@@ -23,7 +26,7 @@ else:
     import dask.array as da
 
     def grid(vis, uvw, flags, weights, ref_wave,
-             convolution_filter, nx=1024, ny=1024):
+             convolution_filter, cell_size, nx=1024, ny=1024):
         """ Documentation below """
 
         # Creation correlation dimension strings for each correlation
@@ -34,6 +37,7 @@ else:
         def _grid_fn(vis, uvw, flags, weights, ref_wave, convolution_filter):
             return np_grid_fn(vis[0], uvw[0], flags[0], weights[0],
                               ref_wave[0], convolution_filter,
+                              cell_size,
                               nx=nx, ny=ny)[None, :]
 
         # Get grids, stacked by row
@@ -51,8 +55,15 @@ else:
         # Sum grids over the row dimension to produce (ny, nx, corr_1, corr_2)
         return grids.sum(axis=0)
 
-    def degrid(grid, uvw, weights, ref_wave, convolution_filter):
+    def degrid(grid, uvw, weights, ref_wave, convolution_filter, cell_size):
         """ Documentation below """
+
+        grid_flat_corrs = reduce(mul, grid.shape[2:])
+        weight_flat_corrs = reduce(mul, weights.shape[2:])
+
+        assert grid_flat_corrs == weight_flat_corrs
+        assert uvw.shape[0] == weights.shape[0]
+        assert weights.shape[1] == ref_wave.shape[0]
 
         # Creation correlation dimension strings for each correlation
         corrs = tuple('corr-%d' for i in range(len(grid.shape[2:])))
@@ -64,6 +75,7 @@ else:
                             ref_wave, ("chan",),
                             concatenate=True,
                             convolution_filter=convolution_filter,
+                            cell_size=cell_size,
                             dtype=np.complex64)
 
 grid.__doc__ = mod_docs(np_grid_fn.__doc__,